Updated emojipedia.js
Issue : According to the bug, the search was happening only after removing the whole word once searched.
Reason : This was happening because the search was being made from the previous search result.
Solution : I have changed the array from which the search was happening, that is, from the whole data every time, instead of searching from the previous data.