Hangman Game is Developed and tested
This game include to guess the user a word using the hint provided. With each wrong guess, a part of the hangman figure appears. If complete hangman is formed then user loses the game. After completion of game, user can get the new word to guess by clicking reset game.